Move untestable code to shims
