+++ invoke/runners.py
@@ -50,8 +50,8 @@ from .terminals import (
     character_buffered,
     ready_for_reading,
     bytes_to_read,
-    get_short_path_name,
 )
+from .shims import get_short_path_name
 from .util import has_fileno, isatty, ExceptionHandlingThread
 
 if TYPE_CHECKING:
@@ -416,8 +416,7 @@ class Runner:
         # Echo running command (wants to be early to be included in dry-run)
         if self.opts["echo"]:
             self.echo(command)
-        if WINDOWS:
-            self.opts["shell"] = get_short_path_name(self.opts["shell"])
+        self.opts["shell"] = get_short_path_name(self.opts["shell"])
         # Prepare common result args.
         # TODO: I hate this. Needs a deeper separate think about tweaking
         # Runner.generate_result in a way that isn't literally just this same
